+++ packages/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shallot-ui/web
 
+## 0.22.7
+
+### Patch Changes
+
+- Fixes treatment of input error styles
+
 ## 0.22.6
 
 ### Patch Changes